create - add components and app.css
