+++ src/app.css
@@ -1,0 +1,17 @@
+#root {
+  position: relative;
+  display: flex;
+  justify-content: center;
+}
+
+.APP{
+  display: flex;
+  flex-direction: column;
+  border: solid 1px red;
+  box-shadow: 0 10px 10px 0 black;
+  align-items: center;
+  min-width: 320px;
+  width: 100%;
+  height: auto;
+  padding: 30px 50px 0 50px;
+}
