+++ src/components/Footer.jsx
@@ -1,0 +1,7 @@
+export const Footer = () => {
+  return (
+    <footer>
+      <p>&copy; 2024 XingS Happy Thought APP. All rights reserved.</p>
+    </footer>  
+  );
+};
