+++ src/components/Header.jsx
@@ -1,0 +1,8 @@
+export const Header = () => {
+  return (
+    <header>
+      <h1>All Thoughts Have Power!</h1>
+      <h2>But Only Happy One Gain Likes❤️</h2>
+    </header>      
+  );
+};
