+++ src/components/Main.jsx
@@ -1,0 +1,16 @@
+import { useEffect } from "react";
+import { NewThought } from "./NewThought";
+import { Thoughts } from "./Thoughts";
+
+export const Main = () => {
+  useEffect(() => {
+    console.log("Main mount")
+  });
+
+  return (
+    <main>
+      <NewThought />
+      <Thoughts />
+    </main> 
+  );
+};
