+++ src/components/Thoughts.jsx
@@ -1,0 +1,12 @@
+export const Thoughts = () => {
+  return (
+    <div className="thought-container">
+      <p>It is my birthday!!!</p>
+      <div className="like-container">
+        <button className="like-btn">❤️</button>
+        <p className="like-number">x 10</p>
+        <p className="time">30 secons ago</p>
+      </div>
+    </div>  
+  );
+};
