+++ src/index.css
@@ -1,5 +1,4 @@
 :root {
-  margin: 0;
   font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
     "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
     sans-serif;
@@ -10,4 +9,30 @@
 code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
     monospace;
+}
+
+body, html {
+  padding: 0;
+  margin: 0;
+  width: 100%;
+  height: 100vh;
+}
+
+#root {
+  position: relative;
+  display: flex;
+  justify-content: center;
+}
+
+
+.APP{
+  display: flex;
+  flex-direction: column;
+  border: solid 1px red;
+  box-shadow: 0 10px 10px 0 black;
+  align-items: center;
+  min-width: 320px;
+  width: 100%;
+  height: auto;
+  padding: 30px 50px 0 50px;
 }
